move helper functions to helper class
